requirements.txt
torch
huggingface_hub
bitsandbytes
librosa
transformers.zip
